% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCIMIS.R
\name{CIMISweather}
\alias{CIMISweather}
\title{Get data from CIMIS API}
\usage{
CIMISweather(start, end, station_nbr, unitOfMeasure = "M", ...,
  include_qc = FALSE, api_key = getOption("Rcimis_key",
  stop("You need a key.")), .opts = list())
}
\arguments{
\item{start}{The start date, in ISO format (e.g. "YYYY-MM-DD"). Cannot be earlier than 1987-06-7}

\item{end}{The end date}

\item{station_nbr}{The CIMIS assigned statoin number. Can be found using \code{get_station_info}}

\item{unitOfMeasure}{"M" for metric units, "E" for empirical}

\item{...}{Additional arguments passed to the CIMIS API.}

\item{include_qc}{Logical, should the quality control flags be included in the output.}

\item{api_key}{The API key. By default, the function checks \code{getOptions} for the api as "Rcimis_key"}

\item{.opts}{options passed to \code{getForm}}
}
\value{
a data frame, JSON object, or XML object, depending on options passed to getCIMIS
}
\description{
This allows the R user to send a query to the CIMIS API to get various
data.
}
\examples{
#Get the station number for the Davis, CA station
get_station_info("Davis")

#Davis is station #6, and data starts 1982-07-17
ans = CIMISweather(start = "1987-07-17", end = Sys.Date())

}
\author{
Matt Espe and Duncan Temple Lang
}

