% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearestStation.R
\name{getDists}
\alias{getDists}
\title{getDists}
\usage{
getDists(lat, lon, n = 1, FUN = distHaversine, cached = TRUE,
  StnInfo = getStationInfo(cached))
}
\arguments{
\item{lat}{The latitude, in decimal degrees}

\item{lon}{longitude, in decimal degrees}

\item{n}{number of nearest stations to return}

\item{FUN}{the function used to calculate the distance. Passed to the \code{distm} function in the \code{geospheres} package}

\item{cached}{logical, should the cached station information be used, or should a up to the date version be retrieved from CIMIS}

\item{StnInfo}{object providing details of the CIMIS weather stations. This can be retrieved
using \code{getStationInfo} or \code{updateStationInfo}.}
}
\value{
a subset of \code{StnInfo} ordered by distance to the specificed point, with the distance in Km added
}
\description{
Get geographic distance to nearest CIMIS stations
}
\details{
This function gets the nearest n stations from a geographic point.
}
\examples{
#Get the 5 closest stations
getDists(37.8, -121.22, n = 5)

}
\author{
Matt Espe
}
