% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCIMIS.R
\name{CIMISweather}
\alias{CIMISweather}
\title{Get data from CIMIS API}
\usage{
CIMISweather(startyear, endyear, station_nbr, unitOfMeasure = "M", ...,
  include_qc = FALSE, api_key = getOption("Rcimis_key",
  stop("You need a key.")), .opts = list())
}
\arguments{
\item{startyear}{}

\item{endyear}{}

\item{station_nbr}{}

\item{...}{}

\item{include_qc}{}

\item{api_key}{}

\item{.opts}{}
}
\description{
This allows the R user to send a query to the CIMIS API to get various
data.
}
\author{
Matt Espe
}

