% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCIMIS.R
\name{CIMISweather}
\alias{CIMISweather}
\title{Get data from CIMIS API}
\usage{
CIMISweather(startDate, endDate, targets, unitOfMeasure = "M",
  dataItems = character(), prioritizeSCS = NA, ..., includeQC = FALSE,
  appKey = getOption("Rcimis_key", stop("You need a key.")), .opts = list())
}
\arguments{
\item{startDate}{The start date, in ISO format (e.g. "YYYY-MM-DD"). Cannot be earlier than 1987-06-7}

\item{endDate}{The end date in ISO format}

\item{targets}{The CIMIS assigned station number, zip code, address, or Lat/Long (comma separated). See \code{StnInfo} dataset for list of stations.}

\item{unitOfMeasure}{"M" for metric units, "E" for empirical}

\item{dataItems}{optional. If not provided, function will return all variables}

\item{prioritizeSCS}{Should results from Spatial CIMIS (interpolated) be prioritized?}

\item{...}{Additional arguments passed to the CIMIS API.}

\item{includeQC}{Should columns with quality control flags be included in results?}

\item{appKey}{The API key. By default, the function checks \code{getOptions} for the api as "Rcimis_key".  This avoids exposing this private, secure information in scripts and console.}

\item{.opts}{options passed to \code{getForm} to control the RCurl HTTP request.}
}
\value{
a data frame, JSON object, or XML object, depending on options passed to getCIMIS
}
\description{
This allows the R user to send a general query to the CIMIS API to get various
data from the API.
}
\details{
This is a general function to get data from the CIMIS API that takes arbitrary API parameters.
}
\examples{
#Get the station number for the Davis, CA station
# Fix later

#Davis is station #6, and data starts 1982-07-17
ans = CIMISweather(startDate = "1987-07-17", endDate = Sys.Date(), targets = 6)

}
\author{
Matt Espe and Duncan Temple Lang
}

