% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCIMIS.R
\name{getCIMIS}
\alias{getCIMIS}
\title{Get CIMIS data}
\usage{
getCIMIS(start, end, api_key = getOption("Rcimis_key",
  stop("You need a key.")), ..., .opts = list(),
  url = "http://et.water.ca.gov/api/data")
}
\arguments{
\item{start}{start date, in ISO format ("YYYY-mm-dd")}

\item{end}{end date, in ISO format}

\item{api_key}{a personal API key for accessing data}

\item{...}{additional arguments in the form of "key = value" pairs
passed along as query parameters}

\item{url}{default URL to query}
}
\value{
a data.frame object of query results
}
\description{
Query the CIMIS API
}
\details{
.. content for \details{} ..
}
\author{
Matthew Espe
}

