% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCIMIS.R
\name{getCIMIS}
\alias{getCIMIS}
\title{Get CIMIS data}
\usage{
getCIMIS(start, end, targets = NA, unitOfMeasure = NA, dataItems = NA,
  prioritizeSCS = NA, appKey = getOption("Rcimis_key",
  stop("You need a key.")), .opts = list(), format = "json",
  url = "http://et.water.ca.gov/api/data", parseJSON = (format == "json"))
}
\arguments{
\item{start}{start date, in ISO format ("YYYY-mm-dd")}

\item{end}{end date, in ISO format}

\item{targets}{Either the station number, the zip code, address, or lat/long separated by commas. See API documentation for more information}

\item{unitOfMeasure}{"M" for metric units, "E" for empirial, defaults to "E"}

\item{dataItems}{optional, if not provided will return all daily variables}

\item{prioritizeSCS}{should data from Spatial CIMIS (interpolated) be prioritized?}

\item{appKey}{The API key for CIMIS. By default, the function looks in getOption}

\item{.opts}{Additional options passed to \code{getForm}}

\item{format}{XML or JSON, defaults to JSON}

\item{url}{default URL to query}

\item{parseJSON}{logical value indicating whether to process the result as JSON or just return it.
If format is not "json, then this defaults to returning the document from CIMIS.}
}
\value{
a data.frame, XML, or JSON (depending on above options) object of query results
}
\description{
Query the CIMIS API
}
\examples{
## Get Evapo-transpiration for Jan 2015 in Metric units from station 6
getCIMIS("2015-01-01", "2015-02-01", 6, "M", "day-eto", FALSE)


}
\author{
Matthew Espe and Duncan Temple Lang
}

