% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearestStation.R
\name{getDists}
\alias{getDists}
\title{getDists}
\usage{
getDists(lat, lon, n = 1, FUN = distHaversine)
}
\arguments{
\item{lat}{The latitude, in decimal degrees}

\item{lon}{longitude, in decimal degrees}

\item{n}{number of nearest stations to return}

\item{FUN}{the function used to calculate the distance. Passed to the \code{distm} function in the \package{geospheres} package}
}
\value{
a subset of \code{StnInfo} ordered by distance to the specificed point, with the distance in Km added
}
\description{
Get geographic distance to nearest CIMIS stations
}
\details{
This function gets the nearest n stations from a geographic point.
}
\examples{
#Get the 5 closest stations
getDists(37,8, -121.22, n = 5)

}
\author{
Matt Espe
}

